triangle36
[18:29:40] * triangle36 has a crazy idea... there are no dynamics on the
client, and there is a boolean even on server worlds (isWorldRamote or
something) that mite be worth messing with
[18:35:32] triangle36: ok its World.isClientSide (not the Bukkit api
World, but the one you should not mess with)
[18:36:13] * triangle36 is going to set that to true and see what
happens
[18:37:09] triangle36: oops... its final, aka needs a little hacking :/
[18:54:17] triangle36: :O
[18:54:31] triangle36: it worked! sand does not fall when I place it! XD
[18:57:03] * triangle36 set World.isClientWorld to true on a server
world! no side effects... yet
[18:59:49] triangle36: ok I found some side effects, entitys cant be
spawned, even with /summon and tnt placed next to redstone block
vanishes, but redstone block placed next to tnt does nothing
[19:14:13] triangle36: RoboMWM: to summarize, World.isClientSide set to
true on the server side tricks the server into disabling all dynamics.
(its a bit of a hack though)
[19:15:14] triangle36: that boolean is used all over the place form what
I have seen (not just in the chunk classes)
[19:22:21] triangle36: RoboMWM: here is a screenshot of this "config
world" I was working on http://i.imgur.com/JVcqvlD.png (just to satisfy
curiosity)
[19:26:16] triangle36: RoboMWM: how what goes? World.isClientSide set to
true? I did that and it disables dynamics with some miner side-effects
[19:28:11] triangle36: remember when I said /summon does nothing? it was
spawn eggs that did nothing... /summon spawns a frozen entity (kinda
cool)
[19:30:03] triangle36: found another side effect! /kill is vary buggy
and does not kill most of the time
[19:30:40] triangle36: (kills ender dragons, but not pigs withers or
players)
[19:39:58] triangle36: those fall under the "miner side-effects" :)
[19:53:19] triangle36: RoboMWM: ok, I will give you part of the code to
do that... one sec
[19:57:29] triangle36: RoboMWM: here is a function to set
World.isClientSide to true for a given world...
http://pastebin.com/m90v2hi0
[20:06:41] * triangle36 wonders if RoboMWM is making a plugin to test
out that code